Extract env setup page and OAuth page into common folder
