+++ common/package.json
@@ -1,14 +1,6 @@
 {
-  "name": "oauth-connect-popup",
+  "name": "common",
   "version": "0.1.0",
-  "private": true,
-  "scripts": {
-    "preinstall": "cd ../common && npm install",
-    "dev": "next dev",
-    "build": "next build",
-    "start": "next start",
-    "lint": "next lint"
-  },
   "dependencies": {
     "@types/node": "^20.6.2",
     "@types/react": "^18.2.21",
