+++ strong-customer-authentication/package.json
@@ -3,6 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
+    "preinstall": "cd ../common && npm install",
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
