+++ strong-customer-authentication/pages/oauth-connect.tsx
@@ -1,0 +1,3 @@
+// Renders a "Connect your Wise account" page
+export { default } from '../../common/components/OAuthConnectPage/OAuthConnectPage';
+export { getServerSideProps } from '../../common/components/OAuthConnectPage/getServerSideProps';
